+++ packages/app/src/components/BlockList/index.tsx
@@ -18,6 +18,7 @@ interface BlockListProps {
   ee: EventEmitter;
   extraCreators?: ActionCreators;
   flowActions?: {};
+  prefix: string;
   showDialog: ShowDialogAction;
   transitions?: boolean;
 }
@@ -41,6 +42,7 @@ export default function BlockList({
   ee,
   extraCreators,
   flowActions,
+  prefix,
   showDialog,
   transitions,
 }: BlockListProps): React.ReactElement {
@@ -87,6 +89,7 @@ export default function BlockList({
         extraCreators={extraCreators}
         flowActions={flowActions}
         pageReady={pageReady}
+        prefix={`${prefix}.${index}`}
         ready={ready}
         showDialog={showDialog}
       />