+++ packages/app/src/components/FlowPage/index.tsx
@@ -18,6 +18,7 @@ export default function FlowPage({
   definition,
   ee,
   page,
+  prefix,
   showDialog,
 }: FlowPageProps): React.ReactElement {
   const history = useHistory();
@@ -95,11 +96,12 @@ export default function FlowPage({
         showDialog,
         extraCreators: {},
         flowActions,
+        prefix,
         pushNotifications,
         ee,
         pageReady: null,
       }),
-    [definition, ee, flowActions, history, page, pushNotifications, showDialog],
+    [definition, ee, flowActions, history, page, prefix, pushNotifications, showDialog],
   );
 
   return (
@@ -111,6 +113,7 @@ export default function FlowPage({
         data={data}
         ee={ee}
         flowActions={flowActions}
+        prefix={`${prefix}.subPages`}
         showDialog={showDialog}
         transitions
       />