+++ packages/app/src/components/Main/index.tsx
@@ -22,7 +22,7 @@ export default function Main(): React.ReactElement {
   }
 
   let defaultPath;
-  const routes = definition.pages.map((page) => {
+  const routes = definition.pages.map((page, index) => {
     const path = `/${[
       normalize(page.name),
       ...(page.parameters || []).map((parameter) => `:${parameter}`),
@@ -34,7 +34,7 @@ export default function Main(): React.ReactElement {
     }
     return (
       <Route key={path} exact path={path}>
-        <Page page={page} />
+        <Page page={page} prefix={`pages.${index}`} />
       </Route>
     );
   });