+++ packages/app/src/components/Page/index.test.tsx
@@ -65,7 +65,7 @@ it('should render the blocks for a page', () => {
       pages: [page],
     },
   });
-  const wrapper = shallow(<Page page={page} />);
+  const wrapper = shallow(<Page page={page} prefix="pages.0" />);
 
   expect(wrapper).toMatchSnapshot();
 });
@@ -99,7 +99,7 @@ it('should render tabs pages', () => {
       pages: [page],
     },
   });
-  const wrapper = shallow(<Page page={page} />);
+  const wrapper = shallow(<Page page={page} prefix="pages.0" />);
 
   expect(wrapper).toMatchSnapshot();
 });
@@ -133,7 +133,7 @@ it('should render flow page', () => {
       pages: [page],
     },
   });
-  const wrapper = shallow(<Page page={page} />);
+  const wrapper = shallow(<Page page={page} prefix="pages.0" />);
 
   expect(wrapper).toMatchSnapshot();
 });